requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gevent==23.9.1
geventhttpclient==2.0.2
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
huggingface-hub==0.20.2
idna==3.6
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-pytriton==0.2.5
opencv-python==4.9.0.80
packaging==23.2
protobuf==3.20.3
pydantic==2.5.3
pydantic_core==2.14.6
python-rapidjson==1.14
PyYAML==6.0.1
pyzmq==23.2.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
sh==1.14.3
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tokenizers==0.15.0
torch==1.13.0
tqdm==4.66.1
transformers==4.36.2
tritonclient==2.33.0
typing-inspect==0.6.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
wrapt==1.16.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
